requirements.txt
Flask
flask-restful
flask-cors
sentence-transformers
pinecone-client
gunicorn
